# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: VS2000 Smart Office — Lottery Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get current system settings
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemSettings"
    post:
      operationId: createSettings
      tags: [settings]
      summary: Create new settings record
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SystemSettingsInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemSettings"
  /settings/time-windows:
    get:
      operationId: listTimeWindows
      tags: [settings]
      summary: List cashier time windows
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimeWindow"
    post:
      operationId: createTimeWindow
      tags: [settings]
      summary: Create a cashier time window
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TimeWindowInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeWindow"
  /settings/time-windows/{id}:
    patch:
      operationId: updateTimeWindow
      tags: [settings]
      summary: Update time window
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TimeWindowUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeWindow"
    delete:
      operationId: deleteTimeWindow
      tags: [settings]
      summary: Delete time window
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /sales:
    get:
      operationId: listSales
      tags: [sales]
      summary: List sales logs
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: writerId
          schema:
            type: string
        - in: query
          name: dateFrom
          schema:
            type: string
        - in: query
          name: dateTo
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SalesLog"
    post:
      operationId: createSale
      tags: [sales]
      summary: Log a sale
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SalesLogInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesLog"
...
  /payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List payments
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: agentId
          schema:
            type: string
        - in: query
          name: dateFrom
          schema:
            type: string
        - in: query
          name: dateTo
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
    post:
      operationId: createPayment
      tags: [payments]
      summary: Record a payment (cashier only, within time window)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
        "403":
          description: Outside cashier time window
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /payments/{id}/void:
    post:
      operationId: voidPayment
      tags: [payments]
      summary: Void a payment
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoidInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
...
  /reports/writer/{writerId}:
    get:
      operationId: getWriterReport
      tags: [reports]
      summary: Writer performance report
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: writerId
          required: true
          schema:
            type: string
        - in: query
          name: dateFrom
          schema:
            type: string
        - in: query
          name: dateTo
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WriterReport"
  /reports/agent/{agentId}:
    get:
      operationId: getAgentReport
      tags: [reports]
      summary: Agent summary report
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: agentId
          required: true
          schema:
            type: string
        - in: query
          name: dateFrom
          schema:
            type: string
        - in: query
          name: dateTo
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentReport"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    User:
      type: object
      required: [id, fullName, email, role, isActive, createdAt]
      properties:
        id:
          type: string
        fullName:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [director, administrator, cashier, gross_entry, wins_entry, agent]
        isActive:
          type: boolean
        createdAt:
          type: string
        lastLogin:
          type: ["string", "null"]
...
    AgentWithUser:
      type: object
      required: [id, userId, agentCode, fullCode, isActive, createdAt]
      properties:
        id:
          type: string
        userId:
          type: string
        agentCode:
          type: string
        fullCode:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
        user:
          $ref: "#/components/schemas/User"
...
    Writer:
      type: object
      required: [id, agentId, writerCode, fullCode, fullName, isActive, createdAt]
      properties:
        id:
          type: string
        agentId:
          type: string
        writerCode:
          type: string
        fullCode:
          type: string
        fullName:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
...
    SystemSettings:
      type: object
      required: [id, commissionPct, reservePct, effectiveDate, updatedAt]
      properties:
        id:
          type: string
        commissionPct:
          type: string
        reservePct:
          type: string
        effectiveDate:
          type: string
        updatedAt:
          type: string
        updatedBy:
          type: string
    SystemSettingsInput:
      type: object
      required: [commissionPct, reservePct, effectiveDate]
      properties:
        commissionPct:
          type: string
        reservePct:
          type: string
        effectiveDate:
          type: string
    TimeWindow:
      type: object
      required: [id, windowOpen, windowClose, isActive, createdAt]
      properties:
        id:
          type: string
        dayOfWeek:
          type: ["integer", "null"]
        windowOpen:
          type: string
        windowClose:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
    TimeWindowInput:
      type: object
      required: [windowOpen, windowClose]
      properties:
        dayOfWeek:
          type: ["integer", "null"]
        windowOpen:
          type: string
        windowClose:
          type: string
        isActive:
          type: boolean
    TimeWindowUpdate:
      type: object
      properties:
        dayOfWeek:
          type: ["integer", "null"]
        windowOpen:
          type: string
        windowClose:
          type: string
        isActive:
          type: boolean
    SalesLog:
      type: object
      required: [id, writerId, gameType, ticketAmount, saleDate, loggedBy, createdAt]
      properties:
        id:
          type: string
        writerId:
          type: string
        gameType:
          type: string
        ticketAmount:
          type: string
        saleDate:
          type: string
        imageUrl:
          type: ["string", "null"]
        loggedBy:
          type: string
        createdAt:
          type: string
    SalesLogInput:
      type: object
      required: [writerId, gameType, ticketAmount, saleDate]
      properties:
        writerId:
          type: string
        gameType:
          type: string
        ticketAmount:
          type: string
        saleDate:
          type: string
        imageUrl:
          type: string
...
    Payment:
      type: object
      required: [id, agentId, cashierId, amount, paymentDate, isVoided, createdAt]
      properties:
        id:
          type: string
        agentId:
          type: string
        cashierId:
          type: string
        amount:
          type: string
        paymentDate:
          type: string
        notes:
          type: ["string", "null"]
        isVoided:
          type: boolean
        voidedBy:
          type: ["string", "null"]
        voidedReason:
          type: ["string", "null"]
        createdAt:
          type: string
    PaymentInput:
      type: object
      required: [agentId, amount, paymentDate]
      properties:
        agentId:
          type: string
        amount:
          type: string
        paymentDate:
          type: string
        notes:
          type: string
    VoidInput:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
...
    WriterReportRow:
      type: object
      required: [grossSales, commissionAmount, netGross, winsAmount, reserveAmount, writerBalance]
      properties:
        calcDate:
          type: string
        grossSales:
          type: string
        commissionAmount:
          type: string
        netGross:
          type: string
        winsAmount:
          type: string
        reserveAmount:
          type: string
        writerBalance:
          type: string
    WriterReport:
      type: object
      required: [writer, totals, rows]
      properties:
        writer:
          $ref: "#/components/schemas/Writer"
        totals:
          $ref: "#/components/schemas/WriterReportRow"
        rows:
          type: array
          items:
            $ref: "#/components/schemas/WriterReportRow"
    AgentWriterSummary:
      type: object
      required: [writer, totals]
      properties:
        writer:
          $ref: "#/components/schemas/Writer"
        totals:
          $ref: "#/components/schemas/WriterReportRow"
    AgentReport:
      type: object
      required: [agent, totals, writers]
      properties:
        agent:
          $ref: "#/components/schemas/AgentWithUser"
        totals:
          $ref: "#/components/schemas/WriterReportRow"
        writers:
          type: array
          items:
            $ref: "#/components/schemas/AgentWriterSummary"
...